# src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "<h2> WILL BE UPDATED!</h2></br><b>Incomplete</b> API for car workshop system (lacks employees and parts management).</br><a href=\"https://app.swaggerhub.com/apis/soft_eng/project/0.1.1\">Swagger project (contains models)</a>.",
    "version" : "0.1.1",
    "title" : "Car Workshop",
    "contact" : {
      "email" : "k.baciejowski@gmail.com"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/soft_eng/project/0.1.1",
...
  "schemes" : [ "https" ],
  "paths" : {
    "/client" : {
      "post" : {
        "tags" : [ "clients" ],
        "summary" : "Add new client to the system",
        "operationId" : "addClient",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "New client attributes",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/body"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Validation exception"
          }
        }
      },
      "put" : {
        "tags" : [ "clients" ],
        "summary" : "Update an existing client",
        "operationId" : "updateClient",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "New client object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Client"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Validation exception"
          },
          "404" : {
            "description" : "Client not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "write:clients" ]
        } ]
      }
    },
    "/client/{clientId}" : {
      "get" : {
        "tags" : [ "clients" ],
        "summary" : "Find client by ID",
        "description" : "Returns a single client",
        "operationId" : "getClientById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "description" : "Id of client to return",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/Client"
            }
          },
          "400" : {
            "description" : "Invalid Id supplied"
          },
          "404" : {
            "description" : "Client not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "read:clients" ]
        } ]
      }
    },
...
    "/client/{clientId}/set_discount" : {
      "put" : {
        "tags" : [ "clients", "employees" ],
        "summary" : "Set client's discount",
        "description" : "Accepts discount in percentage (0-30)",
        "operationId" : "setDiscount",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "description" : "Client id",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "New client's discount",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/body_2"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Validation exception"
          },
          "404" : {
            "description" : "Client not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "write:clients" ]
        } ]
      }
    },
...
    "/client/{clientId}/visits" : {
      "get" : {
        "tags" : [ "clients", "visits" ],
        "summary" : "Get all client's visits",
        "operationId" : "getClientVisits",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "clientId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Visit"
              }
            }
          },
          "404" : {
            "description" : "Client not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "read:visits" ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "workshop_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "https://example.com/oauth/dialog",
      "flow" : "implicit",
      "scopes" : {
        "write:clients" : "modify clients",
        "read:clients" : "read clients",
        "write:cars" : "modify cars",
        "read:cars" : "read cars",
        "write:visits" : "modify visits",
        "read:visits" : "read visits",
        "write:payments" : "modify payments",
        "read:payments" : "read payments",
        "write:employees" : "modify employees",
        "read:employees" : "read employees"
      }
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    }
  },
  "definitions" : {
    "Employee" : {
      "type" : "object",
      "required" : [ "employeeId", "name", "surname" ],
      "properties" : {
        "employeeId" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 13
        },
        "name" : {
          "type" : "string",
          "example" : "Katarzyna"
        },
        "surname" : {
          "type" : "string",
          "example" : "Nowak"
        }
      }
    },
    "Client" : {
      "type" : "object",
      "required" : [ "clientId", "discount", "name", "phoneNumber", "surname" ],
      "properties" : {
        "clientId" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 152
        },
        "name" : {
          "type" : "string",
          "example" : "Jan"
        },
        "surname" : {
          "type" : "string",
          "example" : "Kowalski"
        },
        "phoneNumber" : {
          "type" : "string",
          "example" : "+48 123 456 789"
        },
        "email" : {
          "type" : "string",
          "example" : "kowalski@example.com"
        },
        "discount" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 5
        }
      },
      "example" : {
        "clientId" : 152,
        "phoneNumber" : "+48 123 456 789",
        "surname" : "Kowalski",
        "name" : "Jan",
        "discount" : 5,
        "email" : "kowalski@example.com"
      }
    },
...
    "Visit" : {
      "type" : "object",
      "required" : [ "date", "visitId" ],
      "properties" : {
        "visitId" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 5512
        },
        "date" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2020-12-24T12:00:00+01"
        },
        "price" : {
          "type" : "number",
          "example" : 100.5
        },
        "requiredActions" : {
          "type" : "array",
          "example" : [ "oil change", "breaks checking" ],
          "items" : {
            "type" : "string"
          }
        },
        "status" : {
          "type" : "string",
          "example" : "atService",
          "enum" : [ "atService", "repaired", "checkedIn" ]
        },
        "assignedEmployees" : {
          "type" : "array",
          "example" : [ {
            "employeeId" : 12,
            "name" : "Bolesław",
            "surname" : "Nowak"
          }, {
            "employeeId" : 13,
            "name" : "Katarzyna",
            "surname" : "Nowak"
          } ],
          "items" : {
            "$ref" : "#/definitions/Employee"
          }
        }
      },
      "example" : {
        "date" : "2020-12-24T12:00:00+01",
        "requiredActions" : [ "oil change", "breaks checking" ],
        "visitId" : 5512,
        "price" : 100.5,
        "assignedEmployees" : [ {
          "employeeId" : 12,
          "name" : "Bolesław",
          "surname" : "Nowak"
        }, {
          "employeeId" : 13,
          "name" : "Katarzyna",
          "surname" : "Nowak"
        } ],
        "status" : "atService"
      }
    },
...
    "body" : {
      "type" : "object",
      "required" : [ "name", "surname" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "example" : "Jan"
        },
        "surname" : {
          "type" : "string",
          "example" : "Kowalski"
        },
        "phoneNumber" : {
          "type" : "string",
          "example" : "+48 123 456 789"
        },
        "email" : {
          "type" : "string",
          "example" : "kowalski@example.com"
        }
      }
    },
...
    "body_2" : {
      "properties" : {
        "discount" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 15
        }
      }
    },
...
  }
}